Relocate session reducer tests
