+++ src/ducks/modules/session/__tests__/session.test.js
@@ -1,7 +1,7 @@
 /* eslint-env jest */
 
 import reducer, { actionCreators, actionTypes } from '../stage';
-import { actionTypes as protocolActionTypes } from '../protocol';
+import { actionTypes as protocolActionTypes } from '../../protocol';
 
 describe('session reducer', () => {
   it('should return the initial state', () => {
